Add libGit2Sharp + Minor change on list

+++ GoM/Program.cs
@@ -6,7 +6,7 @@ using Newtonsoft.Json;
 using System.Linq;
 using System.Text;
 using GoM.Core.Persistence;
-
+using LibGit2Sharp;
 namespace GoM
 {
     class Program
@@ -121,14 +121,17 @@ namespace GoM
                     {
                         ProcessFile(projectPath, fileList);
                     }
-                    else if (Directory.Exists(projectPath))
+                    else if (Directory.Exists(projectPath) && Repository.IsValid(projectPath))
                     {
-                        FileTree ft = new FileTree();
-                        ft.Nodes = new List<FileTree>();
-                        GetNodes(projectPath, ft);
-                        string json = JsonConvert.SerializeObject(ft, Formatting.Indented);
-                        File.WriteAllText(Path.Combine(Directory.GetCurrentDirectory(), "fileList.json"), json);
-                        ProcessDirectory(projectPath, fileList);
+                        
+                            FileTree ft = new FileTree();
+                            ft.Nodes = new List<FileTree>();
+                            GetNodes(projectPath, ft);
+                            string json = JsonConvert.SerializeObject(ft, Formatting.Indented);
+                            //Console.WriteLine(json);
+                            File.WriteAllText(Path.Combine(Directory.GetCurrentDirectory(), "fileList.json"), json);
+                            ProcessDirectory(projectPath, fileList);
+                        
                     }
                     else
                     {
@@ -149,6 +152,7 @@ namespace GoM
             }
             else if (Directory.Exists(path))
             {
+               
                 GetFiles(path, ft);
                 ft.Data = path;
                 foreach (string item in Directory.GetDirectories(path))
@@ -166,8 +170,7 @@ namespace GoM
         public static void GetChildren(string path, FileTree ft)
         {
            if (Directory.Exists(path))
-            {
-               
+            {                
                 foreach (string item in Directory.GetDirectories(path))
                 {
                     FileTree n = new FileTree();
